+++ pages/dao/[symbol]/proposal/components/instructions/raydium/RemoveLiquidity.tsx
@@ -9,11 +9,9 @@ import {
   UiInstruction,
 } from '@utils/uiTypes/proposalCreationTypes'
 import { NewProposalContext } from '../../../new'
-import useGovernanceAssets from '@hooks/useGovernanceAssets'
 import useWalletStore from 'stores/useWalletStore'
 import Input from '@components/inputs/Input'
 import { debounce } from '@utils/debounce'
-import { GovernedMultiTypeAccount } from '@utils/tokens'
 import Select from '@components/inputs/Select'
 import {
   ProgramAccount,
@@ -27,6 +25,7 @@ import { createRemoveLiquidityInstruction } from '@tools/sdk/raydium/createRemov
 import BigNumber from 'bignumber.js'
 import { jsonInfo2PoolKeys } from '@raydium-io/raydium-sdk'
 import { notify } from '@utils/notifications'
+import useGovernedMultiTypeAccounts from '@hooks/useGovernedMultiTypeAccounts'
 
 const RemoveLiquidityRaydium = ({
   index,
@@ -38,14 +37,7 @@ const RemoveLiquidityRaydium = ({
   const connection = useWalletStore((s) => s.connection)
   const wallet = useWalletStore((s) => s.current)
   const { realmInfo } = useRealm()
-  const [governedAccounts, setGovernedAccounts] = useState<
-    GovernedMultiTypeAccount[]
-  >([])
-  const { getGovernedMultiTypeAccounts } = useGovernanceAssets()
-
-  useEffect(() => {
-    getGovernedMultiTypeAccounts().then(setGovernedAccounts)
-  }, [])
+  const { governedMultiTypeAccounts } = useGovernedMultiTypeAccounts()
 
   const shouldBeGoverned = index !== 0 && governance
   const programId: PublicKey | undefined = realmInfo?.programId
@@ -170,7 +162,7 @@ const RemoveLiquidityRaydium = ({
     <>
       <GovernedAccountSelect
         label="Governance"
-        governedAccounts={governedAccounts as GovernedMultiTypeAccount[]}
+        governedAccounts={governedMultiTypeAccounts}
         onChange={(value) => {
           handleSetForm({ value, propertyName: 'governedAccount' })
         }}
