+++ pages/dao/[symbol]/proposal/components/instructions/solend/DepositReserveLiquidityAndObligationCollateral.tsx
@@ -9,9 +9,7 @@ import {
   DepositReserveLiquidityAndObligationCollateralForm,
 } from '@utils/uiTypes/proposalCreationTypes'
 import { NewProposalContext } from '../../../new'
-import useGovernanceAssets from '@hooks/useGovernanceAssets'
 import useWalletStore from 'stores/useWalletStore'
-import { GovernedMultiTypeAccount } from '@utils/tokens'
 import {
   ProgramAccount,
   serializeInstructionToBase64,
@@ -26,6 +24,7 @@ import SolendConfiguration from '@tools/sdk/solend/configuration'
 
 import BigNumber from 'bignumber.js'
 import { BN } from '@project-serum/anchor'
+import useGovernedMultiTypeAccounts from '@hooks/useGovernedMultiTypeAccounts'
 
 const DepositReserveLiquidityAndObligationCollateral = ({
   index,
@@ -37,19 +36,12 @@ const DepositReserveLiquidityAndObligationCollateral = ({
   const connection = useWalletStore((s) => s.connection)
   const wallet = useWalletStore((s) => s.current)
   const { realmInfo } = useRealm()
-  const [governedAccounts, setGovernedAccounts] = useState<
-    GovernedMultiTypeAccount[]
-  >([])
-  const { getGovernedMultiTypeAccounts } = useGovernanceAssets()
+  const { governedMultiTypeAccounts } = useGovernedMultiTypeAccounts()
 
   // Hardcoded gate used to be clear about what cluster is supported for now
   if (connection.cluster !== 'mainnet') {
     return <>This instruction does not support {connection.cluster}</>
   }
-
-  useEffect(() => {
-    getGovernedMultiTypeAccounts().then(setGovernedAccounts)
-  }, [])
 
   const shouldBeGoverned = index !== 0 && governance
   const programId: PublicKey | undefined = realmInfo?.programId
@@ -144,7 +136,7 @@ const DepositReserveLiquidityAndObligationCollateral = ({
     <>
       <GovernedAccountSelect
         label="Governance"
-        governedAccounts={governedAccounts}
+        governedAccounts={governedMultiTypeAccounts}
         onChange={(value) => {
           handleSetForm({ value, propertyName: 'governedAccount' })
         }}
