+++ pages/dao/[symbol]/proposal/components/instructions/solend/WithdrawObligationCollateralAndRedeemReserveLiquidity.tsx
@@ -9,9 +9,7 @@ import {
   WithdrawObligationCollateralAndRedeemReserveLiquidityForm,
 } from '@utils/uiTypes/proposalCreationTypes'
 import { NewProposalContext } from '../../../new'
-import useGovernanceAssets from '@hooks/useGovernanceAssets'
 import useWalletStore from 'stores/useWalletStore'
-import { GovernedMultiTypeAccount } from '@utils/tokens'
 import {
   ProgramAccount,
   serializeInstructionToBase64,
@@ -24,6 +22,7 @@ import SolendConfiguration from '@tools/sdk/solend/configuration'
 import { withdrawObligationCollateralAndRedeemReserveLiquidity } from '@tools/sdk/solend/withdrawObligationCollateralAndRedeemReserveLiquidity'
 import BigNumber from 'bignumber.js'
 import { BN } from '@project-serum/anchor'
+import useGovernedMultiTypeAccounts from '@hooks/useGovernedMultiTypeAccounts'
 
 const WithdrawObligationCollateralAndRedeemReserveLiquidity = ({
   index,
@@ -35,20 +34,12 @@ const WithdrawObligationCollateralAndRedeemReserveLiquidity = ({
   const connection = useWalletStore((s) => s.connection)
   const wallet = useWalletStore((s) => s.current)
   const { realmInfo } = useRealm()
-  const [governedAccounts, setGovernedAccounts] = useState<
-    GovernedMultiTypeAccount[]
-  >([])
-
-  const { getGovernedMultiTypeAccounts } = useGovernanceAssets()
+  const { governedMultiTypeAccounts } = useGovernedMultiTypeAccounts()
 
   // Hardcoded gate used to be clear about what cluster is supported for now
   if (connection.cluster !== 'mainnet') {
     return <>This instruction does not support {connection.cluster}</>
   }
-
-  useEffect(() => {
-    getGovernedMultiTypeAccounts().then(setGovernedAccounts)
-  }, [])
 
   const shouldBeGoverned = index !== 0 && governance
   const programId: PublicKey | undefined = realmInfo?.programId
@@ -143,7 +134,7 @@ const WithdrawObligationCollateralAndRedeemReserveLiquidity = ({
     <>
       <GovernedAccountSelect
         label="Governance"
-        governedAccounts={governedAccounts}
+        governedAccounts={governedMultiTypeAccounts}
         onChange={(value) => {
           handleSetForm({ value, propertyName: 'governedAccount' })
         }}
